feat: move bredcrumbs building logic to useBreadcrumbs
