+++ client-app/shared/catalog/composables/index.ts
@@ -1,2 +1,3 @@
 export { default as useCategories } from "./useCategories";
 export { default as useProducts } from "./useProducts";
+export { default as useBreadcrumbs } from "./useBreadcrumbs";
